+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 ## Changelog
+* 2024-09-25
+	- LM styled `analysis3.m`.
+	- Changed citation title.
+	- Added data corresponding to `analysis3.m` to data repository.
+* 2024-09-16
+	- AB added spectrogram analysis steps to `analysis3.m` using cosine-bell windows.
+* 2023-11-30
+	- LM added `analysis3.m` spectrogram analysis using multi-tapers method.
 * 2023-11-10
 	- Updated code to work with FPA v2.0.0
 * 2023-09-01